requirements_app.txt
azure-storage-blob==12.8.1
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
blis==0.7.4
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cycler==0.10.0
cymem==2.0.5
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
entrypoints==0.3
Flask==2.0.2
idna==3.3
imutils==0.5.4
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nbstripout==0.5.0
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.3
opencv-python==4.5.4.58
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pathy==0.6.0
pickleshare==0.7.5
Pillow==8.4.0
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.20
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.3
pywin32==302
pywinpty==1.1.4
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests==2.26.0
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
smart-open==5.2.1
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.1
terminado==0.12.1
testpath==0.5.0
thinc==8.0.11
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
